refactor: move run_campaign.py and run_iteration.py into orchestrator package
Enables proper pip install without editable mode. The nous CLI
can now import orchestrator.campaign and orchestrator.iteration
as regular package modules.

Thin wrapper scripts remain at root for backward compatibility.
